Revert "set up the mini version (#4)"
This reverts commit 56a196ffef712a39684705e5474f29efc3d6bc1c.
