--- .env.example
@@ -1,2 +1,0 @@
-# Copy `.env.example` to `.env.local` and modify the value accordingly.
-OPENAI_API_KEY=''
