--- README.md
@@ -1,29 +1,0 @@
-# Mini Builder
-
-![screenshot](./CleanShot%202023-12-14%20at%2015.27.04@2x.jpg)
-
-## Overview
-
-Mini Builder is a streamlined development tool that includes a suite of features to enhance your productivity:
-
-1. **Workflow Editor**: A developer-friendly interface to create and modify workflows.
-1. **Workflow Debugger**: An integrated debugging tool to troubleshoot workflows in real-time.
-1. **Workflow Steps**: Define and manage the sequence of operations within your workflow.
-1. **Workflow Variables**: Manage and utilize variables throughout your workflow processes.
-
-## Key Components
-
-Mini Builder is primarily composed of two essential files:
-
-1. `/app/page.tsx` is the frontend code, which encompasses the workflow editor and debugger interfaces.
-1. `/app/api/chat/route.ts` is the backend code, which includes the workflow engine responsible for executing the defined workflows.
-
-## Getting Started
-
-To get Mini Builder up and running on your local machine, follow these steps:
-
-1. Set up your environment variables based on the provided `.env.example` file.
-1. Install all the necessary dependencies by running `npm install`.
-1. Kickstart the development server with `npm run dev`.
-
-By following the above steps, you’ll be ready to build, edit, and debug workflows in no time!
