+++ app/api/chat/route.ts
@@ -1,67 +1,50 @@
+import { revalidatePath } from 'next/cache'
 import { NextRequest, NextResponse } from 'next/server'
+import { kv } from '@vercel/kv'
 import { OpenAIStream, StreamingTextResponse } from 'ai'
 import OpenAI from 'openai'
 
-import { IConfig } from '@/app/config'
+import { getBot } from '@/lib/actions/bot'
 
 export const runtime = 'edge'
 
+const openai = new OpenAI({
+  apiKey: process.env.OPENAI_API_KEY,
+  baseURL: process.env.OPENAI_BASE_PATH,
+})
+
+// gpt-3.5-turbo-1106
+// gpt-4-1106-preview
 const model = 'gpt-4-1106-preview'
 
 export async function POST(req: NextRequest) {
   try {
-    const { config, messages: _messages } = await req.json()
+    const { code, id, messages: _messages, botId } = await req.json()
 
-    const openai = new OpenAI({
-      apiKey: process.env.OPENAI_API_KEY,
-    })
+    const isValid = process.env.CODE?.split(',').includes(code)
+    if (!isValid) {
+      return NextResponse.json(
+        { ok: false, message: 'Invalid code' },
+        { status: 403 }
+      )
+    }
 
-    const { variables, workflow } = config as IConfig
+    let instruction =
+      'You are ChatGPT, a large language model trained by OpenAI.'
 
-    const systemMessages = []
-    let enabledJsonMode = false
+    if (botId) {
+      const bot = await getBot(botId)
+      if (bot) instruction = bot.instruction
+    }
 
-    const currentStep = workflow.find((w) => !w.done)
-    if (currentStep && currentStep.system_prompt) {
-      let instruction = `${currentStep.system_prompt}`
+    console.log('instruction:', instruction)
 
-      if (instruction.includes('{required_variables}')) {
-        let variables_text = ''
-        if (currentStep.required_variables) {
-          const required_variables = variables.filter(
-            (v) => currentStep.required_variables?.includes(v.key)
-          )
-          variables_text += `${JSON.stringify(required_variables)}`
-        }
-        if (variables_text) {
-          instruction = instruction.replace(
-            '{required_variables}',
-            variables_text
-          )
-        }
-      }
-
-      for (const v of variables) {
-        const key = `{${v.key}}`
-        const find = instruction.includes(key)
-        if (find) {
-          instruction = instruction.replaceAll(key, JSON.stringify(v.value))
-        }
-      }
-
-      systemMessages.push({
+    const systemMessages = [
+      {
         role: 'system',
         content: instruction,
-      })
-      enabledJsonMode = true
-    }
-
-    if (systemMessages.length === 0) {
-      systemMessages.push({
-        role: 'system',
-        content: 'You are ChatGPT, a large language model trained by OpenAI.',
-      })
-    }
+      },
+    ]
 
     const messages = [...systemMessages, ..._messages]
 
@@ -69,12 +52,24 @@ export async function POST(req: NextRequest) {
       model,
       stream: true,
       messages,
-      response_format: {
-        type: enabledJsonMode ? 'json_object' : 'text',
-      },
     })
 
-    const stream = OpenAIStream(response)
+    const stream = OpenAIStream(response, {
+      async onCompletion(completion) {
+        const completionMessage = [
+          {
+            role: 'assistant',
+            content: completion,
+          },
+        ]
+        const messages = [..._messages, ...completionMessage]
+        await kv.hset(`chat:${id}`, {
+          id,
+          messages,
+        })
+        if (botId) revalidatePath(`/b/${botId}`)
+      },
+    })
 
     return new StreamingTextResponse(stream)
   } catch (error: any) {
