+++ app/workflow/config.ts
@@ -1,20 +1,4 @@
-export interface IConfig {
-  version: number
-  variables: IVariable[]
-  workflow: IStep[]
-}
-
-export interface IVariable {
-  key: string
-  description: string
-  value: any
-}
-
-export interface IStep {
-  system_prompt: string
-  required_variables?: string[]
-  done?: boolean
-}
+import { IConfig } from '@/lib/types'
 
 const config: IConfig = {
   version: 0,
