+++ lib/types/bot.ts
@@ -1,0 +1,5 @@
+export interface Bot extends Record<string, any> {
+  id: string
+  name: string
+  instruction: string
+}
