+++ lib/types/chat.ts
@@ -1,0 +1,7 @@
+import { type Message } from 'ai'
+
+export interface Chat extends Record<string, any> {
+  id: string
+  messages: Message[]
+  summary?: string
+}
