+++ lib/types/index.ts
@@ -1,0 +1,3 @@
+export * from './chat'
+export * from './bot'
+export * from './config'
